Move raster preview endpoint into its own router file
